+++ lib/app/_app_view.dart
@@ -1,8 +1,17 @@
 ///
-export 'view/myapp_view.dart';
-
-///
 export 'view/color_picker.dart';
 
 ///
+export 'view/dev_settings.dart';
+
+///
+export 'view/myapp_view.dart';
+
+///
+export 'view/error_widget.dart';
+
+///
 export 'view/google_fonts_spinner.dart';
+
+///
+export 'view/splash_screen.dart';