+++ lib/app/controller/error_controller.dart
@@ -1,0 +1,8 @@
+import '../../_view.dart';
+
+mixin handlesErrors {
+  /// Record the current Error widget builder
+  late ErrorWidgetBuilder errorWidgetBuilder;
+
+  late bool errorInBuild;
+}